// apps/venture/src/plugins/store/e2e/types/generated-admin-types.ts
export type Exact<T extends { [key: string]: unknown }> = {
	[K in keyof T]: T[K];
};
// ...
export type Store = Node & {
	__typename?: 'Store';
	id: Scalars['ID'];
	createdAt: Scalars['DateTime'];
	updatedAt: Scalars['DateTime'];
	name: Scalars['String'];
};

export type StoreList = PaginatedList & {
	__typename?: 'StoreList';
	items: Array<Store>;
	totalItems: Scalars['Int'];
};

export type CreateStoreInput = {
	name: Scalars['String'];
};

export type UpdateStoreInput = {
	id: Scalars['ID'];
	name: Scalars['String'];
};
// ...
export namespace CreateStore {
	export type Variables = CreateStoreMutationVariables;
	export type Mutation = CreateStoreMutation;
	export type CreateStore = NonNullable<CreateStoreMutation['createStore']>;
}

export namespace UpdateStore {
	export type Variables = UpdateStoreMutationVariables;
	export type Mutation = UpdateStoreMutation;
	export type UpdateStore = NonNullable<UpdateStoreMutation['updateStore']>;
}

export type CreateStoreMutationVariables = Exact<{
	input: CreateStoreInput;
}>;

export type CreateStoreMutation = { __typename?: 'Mutation' } & {
	createStore: { __typename?: 'Store' } & Pick<Store, 'id' | 'name'>;
};

export type UpdateStoreMutationVariables = Exact<{
	input: UpdateStoreInput;
}>;

export type UpdateStoreMutation = { __typename?: 'Mutation' } & {
	updateStore: { __typename?: 'Store' } & Pick<Store, 'id' | 'name'>;
};