// apps/venture/src/plugins/store/e2e/types/generated-shop-types.ts
export type Maybe<T> = T | null;
export type Exact<T extends { [key: string]: unknown }> = {
	[K in keyof T]: T[K];
};
export type MakeOptional<T, K extends keyof T> = Omit<T, K> & {
	[SubKey in K]?: Maybe<T[SubKey]>;
};
export type MakeMaybe<T, K extends keyof T> = Omit<T, K> & {
	[SubKey in K]: Maybe<T[SubKey]>;
};
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
	ID: string;
	String: string;
	Boolean: boolean;
	Int: number;
	Float: number;
	/** The `JSON` scalar type represents JSON values as specified by [ECMA-404](http://www.ecma-international.org/publications/files/ECMA-ST/ECMA-404.pdf). */
	JSON: any;
	/** A date-time string at UTC, such as 2007-12-03T10:15:30Z, compliant with the `date-time` format outlined in section 5.6 of the RFC 3339 profile of the ISO 8601 standard for representation of dates and times using the Gregorian calendar. */
	DateTime: any;
	/** The `Upload` scalar type represents a file upload. */
	Upload: any;
};

export type Query = {
	__typename?: 'Query';
	/** The active Channel */
	activeChannel: Channel;
	/** The active Customer */
	activeCustomer?: Maybe<Customer>;
	/**
	 * The active Order. Will be `null` until an Order is created via `addItemToOrder`. Once an Order reaches the
	 * state of `PaymentApproved` or `PaymentSettled`, then that Order is no longer considered "active" and this
	 * query will once again return `null`.
	 */
	activeOrder?: Maybe<Order>;
	/** An array of supported Countries */
	availableCountries: Array<Country>;
	/** A list of Collections available to the shop */
	collections: CollectionList;
	/** Returns a Collection either by its id or slug. If neither 'id' nor 'slug' is speicified, an error will result. */
	collection?: Maybe<Collection>;
	/** Returns a list of eligible shipping methods based on the current active Order */
	eligibleShippingMethods: Array<ShippingMethodQuote>;
	/** Returns a list of payment methods and their eligibility based on the current active Order */
	eligiblePaymentMethods: Array<PaymentMethodQuote>;
	/** Returns information about the current authenticated User */
	me?: Maybe<CurrentUser>;
	/** Returns the possible next states that the activeOrder can transition to */
	nextOrderStates: Array<Scalars['String']>;
	/**
	 * Returns an Order based on the id. Note that in the Shop API, only orders belonging to the
	 * currently-authenticated User may be queried.
	 */
	order?: Maybe<Order>;
	/**
	 * Returns an Order based on the order `code`. For guest Orders (i.e. Orders placed by non-authenticated Customers)
	 * this query will only return the Order within 2 hours of the Order being placed. This allows an Order confirmation
	 * screen to be shown immediately after completion of a guest checkout, yet prevents security risks of allowing
	 * general anonymous access to Order data.
	 */
	orderByCode?: Maybe<Order>;
	/** Get a Product either by id or slug. If neither 'id' nor 'slug' is speicified, an error will result. */
	product?: Maybe<Product>;
	/** Get a list of Products */
	products: ProductList;
	/** Search Products based on the criteria set by the `SearchInput` */
	search: SearchResponse;
	stores: StoreList;
	store?: Maybe<Store>;
};
// ...
export type CurrentUser = {
	__typename?: 'CurrentUser';
	id: Scalars['ID'];
	identifier: Scalars['String'];
	channels: Array<CurrentUserChannel>;
};

export type CurrentUserChannel = {
	__typename?: 'CurrentUserChannel';
	id: Scalars['ID'];
	token: Scalars['String'];
	code: Scalars['String'];
	permissions: Array<Permission>;
};

export type Channel = Node & {
	__typename?: 'Channel';
	id: Scalars['ID'];
	createdAt: Scalars['DateTime'];
	updatedAt: Scalars['DateTime'];
	code: Scalars['String'];
	token: Scalars['String'];
	defaultTaxZone?: Maybe<Zone>;
	defaultShippingZone?: Maybe<Zone>;
	defaultLanguageCode: LanguageCode;
	currencyCode: CurrencyCode;
	pricesIncludeTax: Scalars['Boolean'];
	customFields?: Maybe<Scalars['JSON']>;
};

export type Collection = Node & {
	__typename?: 'Collection';
	id: Scalars['ID'];
	createdAt: Scalars['DateTime'];
	updatedAt: Scalars['DateTime'];
	languageCode?: Maybe<LanguageCode>;
	name: Scalars['String'];
	slug: Scalars['String'];
	breadcrumbs: Array<CollectionBreadcrumb>;
	position: Scalars['Int'];
	description: Scalars['String'];
	featuredAsset?: Maybe<Asset>;
	assets: Array<Asset>;
	parent?: Maybe<Collection>;
	children?: Maybe<Array<Collection>>;
	filters: Array<ConfigurableOperation>;
	translations: Array<CollectionTranslation>;
	productVariants: ProductVariantList;
	customFields?: Maybe<Scalars['JSON']>;
};
// ...
export type CollectionList = PaginatedList & {
	__typename?: 'CollectionList';
	items: Array<Collection>;
	totalItems: Scalars['Int'];
};
// ...
export type Node = {
	id: Scalars['ID'];
};
// ...
export type ShippingMethodQuote = {
	__typename?: 'ShippingMethodQuote';
	id: Scalars['ID'];
	price: Scalars['Int'];
	priceWithTax: Scalars['Int'];
	code: Scalars['String'];
	name: Scalars['String'];
	description: Scalars['String'];
	/** Any optional metadata returned by the ShippingCalculator in the ShippingCalculationResult */
	metadata?: Maybe<Scalars['JSON']>;
};

export type PaymentMethodQuote = {
	__typename?: 'PaymentMethodQuote';
	id: Scalars['ID'];
	code: Scalars['String'];
	name: Scalars['String'];
	description: Scalars['String'];
	isEligible: Scalars['Boolean'];
	eligibilityMessage?: Maybe<Scalars['String']>;
};

export type Country = Node & {
	__typename?: 'Country';
	id: Scalars['ID'];
	createdAt: Scalars['DateTime'];
	updatedAt: Scalars['DateTime'];
	languageCode: LanguageCode;
	code: Scalars['String'];
	name: Scalars['String'];
	enabled: Scalars['Boolean'];
	translations: Array<CountryTranslation>;
};
// ...
export type Customer = Node & {
	__typename?: 'Customer';
	id: Scalars['ID'];
	createdAt: Scalars['DateTime'];
	updatedAt: Scalars['DateTime'];
	title?: Maybe<Scalars['String']>;
	firstName: Scalars['String'];
	lastName: Scalars['String'];
	phoneNumber?: Maybe<Scalars['String']>;
	emailAddress: Scalars['String'];
	addresses?: Maybe<Array<Address>>;
	orders: OrderList;
	user?: Maybe<User>;
	customFields?: Maybe<Scalars['JSON']>;
};
// ...
export type Order = Node & {
	__typename?: 'Order';
	id: Scalars['ID'];
	createdAt: Scalars['DateTime'];
	updatedAt: Scalars['DateTime'];
	/**
	 * The date & time that the Order was placed, i.e. the Customer
	 * completed the checkout and the Order is no longer "active"
	 */
	orderPlacedAt?: Maybe<Scalars['DateTime']>;
	/** A unique code for the Order */
	code: Scalars['String'];
	state: Scalars['String'];
	/** An order is active as long as the payment process has not been completed */
	active: Scalars['Boolean'];
	customer?: Maybe<Customer>;
	shippingAddress?: Maybe<OrderAddress>;
	billingAddress?: Maybe<OrderAddress>;
	lines: Array<OrderLine>;
	/**
	 * Surcharges are arbitrary modifications to the Order total which are neither
	 * ProductVariants nor discounts resulting from applied Promotions. For store,
	 * one-off discounts based on customer interaction, or surcharges based on payment
	 * methods.
	 */
	surcharges: Array<Surcharge>;
	discounts: Array<Discount>;
	/** An array of all coupon codes applied to the Order */
	couponCodes: Array<Scalars['String']>;
	/** Promotions applied to the order. Only gets populated after the payment process has completed. */
	promotions: Array<Promotion>;
	payments?: Maybe<Array<Payment>>;
	fulfillments?: Maybe<Array<Fulfillment>>;
	totalQuantity: Scalars['Int'];
	/**
	 * The subTotal is the total of all OrderLines in the Order. This figure also includes any Order-level
	 * discounts which have been prorated (proportionally distributed) amongst the OrderItems.
	 * To get a total of all OrderLines which does not account for prorated discounts, use the
	 * sum of `OrderLine.discountedLinePrice` values.
	 */
	subTotal: Scalars['Int'];
	/** Same as subTotal, but inclusive of tax */
	subTotalWithTax: Scalars['Int'];
	currencyCode: CurrencyCode;
	shippingLines: Array<ShippingLine>;
	shipping: Scalars['Int'];
	shippingWithTax: Scalars['Int'];
	/** Equal to subTotal plus shipping */
	total: Scalars['Int'];
	/** The final payable amount. Equal to subTotalWithTax plus shippingWithTax */
	totalWithTax: Scalars['Int'];
	/** A summary of the taxes being applied to this Order */
	taxSummary: Array<OrderTaxSummary>;
	history: HistoryEntryList;
	customFields?: Maybe<Scalars['JSON']>;
};
// ...
export type SearchResponse = {
	__typename?: 'SearchResponse';
	items: Array<SearchResult>;
	totalItems: Scalars['Int'];
	facetValues: Array<FacetValueResult>;
};
// ...
export type Product = Node & {
	__typename?: 'Product';
	id: Scalars['ID'];
	createdAt: Scalars['DateTime'];
	updatedAt: Scalars['DateTime'];
	languageCode: LanguageCode;
	name: Scalars['String'];
	slug: Scalars['String'];
	description: Scalars['String'];
	featuredAsset?: Maybe<Asset>;
	assets: Array<Asset>;
	variants: Array<ProductVariant>;
	optionGroups: Array<ProductOptionGroup>;
	facetValues: Array<FacetValue>;
	translations: Array<ProductTranslation>;
	collections: Array<Collection>;
	customFields?: Maybe<Scalars['JSON']>;
};

export type ProductTranslation = {
	__typename?: 'ProductTranslation';
	id: Scalars['ID'];
	createdAt: Scalars['DateTime'];
	updatedAt: Scalars['DateTime'];
	languageCode: LanguageCode;
	name: Scalars['String'];
	slug: Scalars['String'];
	description: Scalars['String'];
};

export type ProductList = PaginatedList & {
	__typename?: 'ProductList';
	items: Array<Product>;
	totalItems: Scalars['Int'];
};
// ...
export type Store = Node & {
	__typename?: 'Store';
	id: Scalars['ID'];
	createdAt: Scalars['DateTime'];
	updatedAt: Scalars['DateTime'];
	name: Scalars['String'];
};

export type StoreList = PaginatedList & {
	__typename?: 'StoreList';
	items: Array<Store>;
	totalItems: Scalars['Int'];
};
// ...
export namespace GetStores {
	export type Variables = GetStoresQueryVariables;
	export type Query = GetStoresQuery;
	export type Stores = NonNullable<GetStoresQuery['stores']>;
	export type Items = NonNullable<
		NonNullable<NonNullable<GetStoresQuery['stores']>['items']>[number]
	>;
}

export namespace GetStore {
	export type Variables = GetStoreQueryVariables;
	export type Query = GetStoreQuery;
	export type Store = NonNullable<GetStoreQuery['store']>;
}

export type GetStoresQueryVariables = Exact<{ [key: string]: never }>;

export type GetStoresQuery = { __typename?: 'Query' } & {
	stores: { __typename?: 'StoreList' } & Pick<StoreList, 'totalItems'> & {
			items: Array<{ __typename?: 'Store' } & Pick<Store, 'name'>>;
		};
};

export type GetStoreQueryVariables = Exact<{
	id: Scalars['ID'];
}>;

export type GetStoreQuery = { __typename?: 'Query' } & {
	store?: Maybe<{ __typename?: 'Store' } & Pick<Store, 'name'>>;
};